backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
whitenoise>=6.8,<7.0
celery>=5.4,<6.0
redis>=5.2,<6.0
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.3,<3.0
gunicorn>=23.0,<24.0
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
